requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
